Move BeamHasInsertOnConflict instance to Database.Beam.Postgres
