+++ beam-postgres/Database/Beam/Postgres.hs
@@ -1,3 +1,4 @@
+{-# OPTIONS_GHC -fno-warn-orphans #-}
 -- | Postgres is a popular, open-source RDBMS. It is fairly standards compliant
 -- and supports many advanced features and data types.
 --
@@ -67,7 +68,16 @@ module Database.Beam.Postgres
 
   ) where
 
+import Data.Proxy (Proxy(..))
+
+import Database.Beam.Backend
+import Database.Beam.Backend.SQL.BeamExtensions
+import Database.Beam.Query hiding (insert)
+import Database.Beam.Query.Internal
+import Database.Beam.Schema.Tables
+
 import Database.Beam.Postgres.Connection
+import Database.Beam.Postgres.Full
 import Database.Beam.Postgres.Syntax hiding (PostgresInaccessible)
 import Database.Beam.Postgres.Types
 import Database.Beam.Postgres.PgSpecific
@@ -80,3 +90,75 @@ import Database.Beam.Postgres.Extensions ( PgExtensionEntity, IsPgExtension(..)
 import Database.Beam.Postgres.Debug
 
 import qualified Database.PostgreSQL.Simple as Pg
+
+instance BeamHasInsertOnConflict Postgres where
+  type SqlConflictTarget Postgres table = PgInsertOnConflictTarget table
+  type SqlConflictAction Postgres table = PgConflictAction table
+
+  insertOnConflict tbl vs target action = insert tbl vs $ onConflict target action
+
+  -- | Perform the conflict action when any constraint or index conflict occurs.
+  -- Syntactically, this is the @ON CONFLICT@ clause, without any /conflict target/.
+  anyConflict = PgInsertOnConflictTarget (\_ -> PgInsertOnConflictTargetSyntax mempty)
+
+  -- | The Postgres @DO NOTHING@ action
+  onConflictDoNothing = PgConflictAction $ \_ -> PgConflictActionSyntax (emit "DO NOTHING")
+
+  -- | The Postgres @DO UPDATE SET@ action, without the @WHERE@ clause. The
+  -- argument takes an updatable row (like the one used in 'update') and the
+  -- conflicting row. Use 'current_' on the first argument to get the current
+  -- value of the row in the database.
+  onConflictUpdateSet mkAssignments =
+    PgConflictAction $ \tbl ->
+    let QAssignment assignments = mkAssignments tbl tblExcluded
+        tblExcluded = changeBeamRep (\(Columnar' (QField _ _ nm)) -> Columnar' (QExpr (\_ -> fieldE (qualifiedField "excluded" nm)))) tbl
+
+        assignmentSyntaxes =
+          [ fromPgFieldName fieldNm <> emit "=" <> pgParens (fromPgExpression expr)
+          | (fieldNm, expr) <- assignments ]
+    in PgConflictActionSyntax $
+       emit "DO UPDATE SET " <> pgSepBy (emit ", ") assignmentSyntaxes
+
+  -- | The Postgres @DO UPDATE SET@ action, with the @WHERE@ clause. This is like
+  -- 'onConflictUpdateSet', but only rows satisfying the given condition are
+  -- updated. Sometimes this results in more efficient locking. See the Postgres
+  -- <https://www.postgresql.org/docs/current/static/sql-insert.html manual> for
+  -- more information.
+  onConflictUpdateSetWhere mkAssignments where_ =
+    PgConflictAction $ \tbl ->
+    let QAssignment assignments = mkAssignments tbl tblExcluded
+        QExpr where_' = where_ (changeBeamRep (\(Columnar' f) -> Columnar' (current_ f)) tbl)
+        tblExcluded = changeBeamRep (\(Columnar' (QField _ _ nm)) -> Columnar' (QExpr (\_ -> fieldE (qualifiedField "excluded" nm)))) tbl
+
+        assignmentSyntaxes =
+          [ fromPgFieldName fieldNm <> emit "=" <> pgParens (fromPgExpression expr)
+          | (fieldNm, expr) <- assignments ]
+    in PgConflictActionSyntax $
+       emit "DO UPDATE SET " <> pgSepBy (emit ", ") assignmentSyntaxes <> emit " WHERE " <> fromPgExpression (where_' "t")
+
+  -- | Perform the conflict action only when these fields conflict. The first
+  -- argument gets the current row as a table of expressions. Return the conflict
+  -- key. For more information, see the @beam-postgres@ manual.
+  conflictingFields makeProjection =
+    PgInsertOnConflictTarget $ \tbl ->
+    PgInsertOnConflictTargetSyntax $
+    pgParens (pgSepBy (emit ", ") $
+              map fromPgExpression $
+              project (Proxy @Postgres) (makeProjection tbl) "t") <>
+    emit " "
+
+  -- | Like 'conflictingFields', but only perform the action if the condition
+  -- given in the second argument is met. See the postgres
+  -- <https://www.postgresql.org/docs/current/static/sql-insert.html manual> for
+  -- more information.
+  conflictingFieldsWhere makeProjection makeWhere =
+    PgInsertOnConflictTarget $ \tbl ->
+    PgInsertOnConflictTargetSyntax $
+    pgParens (pgSepBy (emit ", ") $
+              map fromPgExpression (project (Proxy @Postgres)
+                                            (makeProjection tbl) "t")) <>
+    emit " WHERE " <>
+    pgParens (let QExpr mkE = makeWhere tbl
+                  PgExpressionSyntax e = mkE "t"
+              in e) <>
+    emit " "
